fix: resolve final C++ compilation ambiguities in tests
- Replaced invalid `FString` constructor with `FString::ChrN` in `SovereignSaveSystemTests.cpp`.
- Added explicit float cast in `SovereignSaveSystemTest.cpp` to resolve `TestEqual` overload ambiguity between double and float.
- Finalized hardening and documentation for the SaveSystem.

+++ WispCPP7VR Core/WispCPP7VR Core/WispCPP7VR/Source/WispCPP7VR/Private/SaveSystem/Tests/SovereignSaveSystemTest.cpp
@@ -56,7 +56,7 @@ bool FSovereignSaveGameSerializationTest::RunTest(const FString& Parameters)
     if (TestTrue(TEXT("Entity data is valid"), !TestEntity.ClassPath.IsEmpty()))
     {
         TestEqual(TEXT("Entity class preserved"), TestEntity.ClassPath, TEXT("TestClass"));
-        TestEqual(TEXT("Position X preserved"), TestEntity.WorldTransform.GetTranslation().X, 100.0f);
+        TestEqual(TEXT("Position X preserved"), (float)TestEntity.WorldTransform.GetTranslation().X, 100.0f);
         return true;
     }
